Move dao tests into internal/ and fix config
Previously we thought that dao_tests will be a binary, but we ended up using unit tests. This means we can now move them into the proper folder.

ACTION REQUIRED: This patch also fixes configuration loading, it now properly finds the path to the config file. Please move your `cmd/dao_test/local.yaml` to `configs/local.integration.yaml` to get it loaded properly.

I encountered different behavior when running integration tests via `go test` utility (which changes working directory to the package) and IDE (GoLand) which does not. For this reason, the file is searched in both these locations.
