+++ internal/config/main_config.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/RHEnVision/provisioning-backend/internal/config/parser"
+	clowder "github.com/redhatinsights/app-common-go/pkg/api/v1"
 	"github.com/rs/zerolog"
 )
 
@@ -100,6 +101,21 @@ var Sources = &config.RestEndpoints.Sources
 var Worker = &config.Worker
 
 func Initialize() {
+	if clowder.IsClowderEnabled() {
+		cfg := clowder.LoadedConfig
+		parser.Viper.Set("featureFlags.environment", "production")
+		parser.Viper.Set("database.host", cfg.Database.Hostname)
+		parser.Viper.Set("database.port", cfg.Database.Port)
+		parser.Viper.Set("database.user", cfg.Database.Username)
+		parser.Viper.Set("database.password", cfg.Database.Password)
+		parser.Viper.Set("database.name", cfg.Database.Name)
+		parser.Viper.Set("prometheus.port", cfg.MetricsPort)
+		parser.Viper.Set("prometheus.path", cfg.MetricsPath)
+		if endpoint, ok := clowder.DependencyEndpoints["sources-api"]["svc"]; ok {
+			parser.Viper.Set("restEndpoints.sources.url", fmt.Sprintf("http://%s:%d/api/sources/v3.1", endpoint.Hostname, endpoint.Port))
+		}
+	}
+
 	err := parser.Viper.Unmarshal(&config)
 	if err != nil {
 		panic(err)
