+++ internal/dao/tests/reservation_test.go
@@ -1,7 +1,7 @@
 //go:build integration
 // +build integration
 
-package main
+package tests
 
 import (
 	"context"
